feat(blewis): Initial commit
initial commit of `blewis`. Whilst the goals of this project are lofty,
I do not intend on doing this quickly.

I aim to complete version 0 in about 6 months. I've written Rust
recreationally quite a bit and started doing so about 2 years ago. Since
then I've not really had the time to play with it anywhere near as much
as I'd have liked.

As of late, I've written a couple small-ish projects just to get back
to grips with the language, and the aim of this project is to be my
magnum-opus.