docs(blewis-considerations): Add CONSIDERATIONS.md
This file is more just a version controlled jot pad where I scribe ideas
as I'm going, to prevent me from getting too distracted from the thing
I'm currently doing. This is not a usual convention but I find it
certainly helps. I can just just add a Harpoon marker for the
considerations file, switch to it in two keypresses, scrawl the tangent
down and press two more keys to get back to exactly where I was in the
file I was currently editing.

This might be helpful when implementing this in other languages too.